Deprecating Mapper\UserInterface. Replaced by Mapper\UserMapperInterface
